VBO textures WORKING WOHO

+++ SpaceShooter/SpaceShooter/MeshXmlLoader.h
@@ -81,7 +81,7 @@ private:
 	bool BindNormals(std::vector<float>& normals, MeshInfo& meshInfo);
 	bool BindColors(std::vector<float>& colors, MeshInfo& meshInfo);
 	bool BindIndices(std::vector<unsigned int>& indices, MeshInfo& meshInfo);
-	bool BindTexCoords(std::vector<unsigned int>& texCoords, MeshInfo& meshInfo);
+	bool BindTexCoords(std::vector<float>& texCoords, MeshInfo& meshInfo);
 
 };
